services/ocr/requirements.txt
fastapi>=0.115.5
uvicorn[standard]>=0.32.0
pydantic>=2.9.2
pydantic-settings>=2.6.1
python-multipart>=0.0.9
openai>=1.53.0
httpx>=0.27.2
minio>=7.2.5
pillow>=11.0.0
gradio>=5.4.0
pymupdf>=1.24.10
